Refactor nvinfer processing.
