+++ savant/gstreamer/element_factory.py
@@ -1,8 +1,11 @@
 """GStreamer pipeline elements factory."""
 
+from typing import Union
+
 from gi.repository import Gst  # noqa:F401
 
-from savant.config.schema import PipelineElement
+from savant.base.model import AttributeModel, ComplexModel, ObjectModel
+from savant.config.schema import ModelElement, PipelineElement
 
 
 class CreateElementException(Exception):
@@ -25,6 +28,9 @@ class GstElementFactory:
         if element.element == 'videotestsrc':
             return self.create_videotestsrc(element)
 
+        if isinstance(element, ModelElement):
+            return self.create_model_element(element)
+
         if isinstance(element, PipelineElement):
             return self.create_element(element)
 
@@ -36,11 +42,17 @@ class GstElementFactory:
     def create_element(element: PipelineElement) -> Gst.Element:
         """Creates Gst.Element.
 
-        :param element: pipeline element to create.
+        :param element: PipelineElement to create.
+        :raises CreateElementException: Unable to create element.
+        :return: Created Gst.Element
         """
         gst_element = Gst.ElementFactory.make(element.element, element.name)
         if not gst_element:
             raise CreateElementException(f'Unable to create element {element}.')
+
+        # set element name from GstElement
+        if element.name is None:
+            element.name = gst_element.name
 
         for prop_name, prop_value in element.properties.items():
             if prop_value is not None:
@@ -49,7 +61,35 @@ class GstElementFactory:
         return gst_element
 
     @staticmethod
+    def create_model_element(element: ModelElement) -> Gst.Element:
+        """Creates Gst.Element for ModelElement.
+
+        :param element: ModelElement to create.
+        :return: Created Gst.Element
+        """
+
+        model: Union[AttributeModel, ComplexModel, ObjectModel] = element.model
+
+        if model.input.preprocess_object_meta:
+            model.input.preprocess_object_meta.load_user_code()
+        if model.input.preprocess_object_image:
+            model.input.preprocess_object_image.load_user_code()
+        if model.output.converter:
+            model.output.converter.load_user_code()
+        if isinstance(model, (ObjectModel, ComplexModel)):
+            for obj in model.output.objects:
+                if obj.selector:
+                    obj.selector.load_user_code()
+
+        return GstElementFactory.create_element(element)
+
+    @staticmethod
     def create_caps_filter(element: PipelineElement) -> Gst.Element:
+        """Creates ``capsfilter`` Gst.Element.
+
+        :param element: Element to create.
+        :return: Created Gst.Element
+        """
         caps = None
         if 'caps' in element.properties and isinstance(element.properties['caps'], str):
             caps = Gst.Caps.from_string(element.properties['caps'])
@@ -61,10 +101,15 @@ class GstElementFactory:
 
     @staticmethod
     def create_videotestsrc(element: PipelineElement) -> Gst.Bin:
-        """videotestsrc element as Gst.Bin with `pad-added`."""
+        """Creates ``videotestsrc`` element as a Gst.Bin with ``pad-added``.
+
+        :param element: Element to create.
+        :return: Created Gst.Element
+        """
+        src_element = GstElementFactory.create_element(element)
+
         src_decodebin = Gst.Bin.new(element.name)
 
-        src_element = GstElementFactory.create_element(element)
         Gst.Bin.add(src_decodebin, src_element)
 
         decodebin = GstElementFactory.create_element(PipelineElement('decodebin'))
